authcontroller service and repo
